Изменение : класс CH264RandomAccess полностью взял у автора LAV, это позволило избавиться от еще одной части кастомного кода на стороне ffmpeg. Так же - в классе CH264Nalu заменил тип int на size_t, т.к. не может быть отрицательных значений.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@2420 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/DSUtil/H264Nalu.h
@@ -45,13 +45,13 @@ private :
 	int			nal_reference_idc;	//! NALU_PRIORITY_xxxx
 	NALU_TYPE	nal_unit_type;		//! NALU_TYPE_xxxx
 
-	int			m_nNALStartPos;		//! NALU start (including startcode / size)
-	int			m_nNALDataPos;		//! Useful part
+	size_t		m_nNALStartPos;		//! NALU start (including startcode / size)
+	size_t		m_nNALDataPos;		//! Useful part
 
 	BYTE*		m_pBuffer;
-	int			m_nCurPos;
-	int			m_nNextRTP;
-	int			m_nSize;
+	size_t		m_nCurPos;
+	size_t		m_nNextRTP;
+	size_t		m_nSize;
 	int			m_nNALSize;
 
 	bool		MoveToNextAnnexBStartcode();
@@ -63,17 +63,17 @@ public :
 	NALU_TYPE GetType() const { return nal_unit_type; };
 	bool IsRefFrame()   const { return (nal_reference_idc != 0); };
 
-	int GetDataLength() const { return m_nCurPos - m_nNALDataPos; };
+	size_t GetDataLength() const { return m_nCurPos - m_nNALDataPos; };
 	BYTE* GetDataBuffer()     { return m_pBuffer + m_nNALDataPos; };
-	int GetRoundedDataLength() const {
-		int nSize = m_nCurPos - m_nNALDataPos;
+	size_t GetRoundedDataLength() const {
+		size_t nSize = m_nCurPos - m_nNALDataPos;
 		return nSize + 128 - (nSize %128);
 	}
 
-	int GetLength() const { return m_nCurPos - m_nNALStartPos; };
+	size_t GetLength() const { return m_nCurPos - m_nNALStartPos; };
 	BYTE* GetNALBuffer()  { return m_pBuffer + m_nNALStartPos; };
 	bool IsEOF() const    { return m_nCurPos >= m_nSize; };
 
-	void SetBuffer(BYTE* pBuffer, int nSize, int nNALSize);
+	void SetBuffer(BYTE* pBuffer, size_t nSize, int nNALSize);
 	bool ReadNext();
 };
